Extract attendee refund routes into attendee-refunds.ts
Move all refund logic (single refund, bulk refund-all) from attendees.ts
into a dedicated attendee-refunds.ts route file:
- refundError, handleAdminAttendeeRefundGet, handleAttendeeRefund
- handleAdminRefundAllGet, handleAdminRefundAllPost, processRefundAll
- getRefundable, refund constants, REFUND_BATCH_LIMIT

Export shared helpers from attendees.ts: attendeeGetRoute,
verifiedAttendeeAction, getReturnUrl, AttendeeWithEvent,
EventRouteParams, NO_PROVIDER_ERROR.
