+++ src/routes/admin/index.ts
@@ -10,6 +10,7 @@
 import { enableQueryLog } from "#lib/db/query-log.ts";
 import { settings } from "#lib/db/settings.ts";
 import { apiKeysRoutes } from "#routes/admin/api-keys.ts";
+import { attendeeRefundRoutes } from "#routes/admin/attendee-refunds.ts";
 import { attendeesRoutes } from "#routes/admin/attendees.ts";
 import { authRoutes } from "#routes/admin/auth.ts";
 import { calendarRoutes } from "#routes/admin/calendar.ts";
@@ -42,6 +43,7 @@ const adminRoutes = {
   ...calendarRoutes,
   ...eventsRoutes,
   ...attendeesRoutes,
+  ...attendeeRefundRoutes,
   ...usersRoutes,
   ...guideRoutes,
   ...groupsRoutes,
